// Frontend/src/app/models/product.model.ts
export interface SanPham {
    maSP: number;
    tenSP: string;
    anhDaiDien: string;
    soLuong: number;
    kichThuoc: string;
    trongLuong: string;
    matKinh: string;
    chatLieuDay: string;
    chongNuoc: string;
    loaiMay: string;
    moTa: string;
    ngayTao: string;
    trangThai: number;
    maDSP: number;

    tenDSP: string;
    giaBan: number;
    phanTramGiamGia: number;
    giaSauGiam: number;
}